feature: 流水线执行终止后，公共构建机的agent进程退出时支持发出kill信号量 #1177
