+++ src/backend/ci/core/worker/worker-agent/src/main/kotlin/com/tencent/devops/agent/utils/KillBuildProcessTree.kt
@@ -27,7 +27,6 @@
 package com.tencent.devops.agent.utils
 
 import com.tencent.devops.common.api.enums.OSType
-import com.tencent.devops.dispatch.pojo.thirdPartyAgent.AgentBuildBaseInfo
 import com.tencent.devops.worker.common.env.AgentEnv
 import com.tencent.process.ProcessTree
 import org.slf4j.LoggerFactory
@@ -59,20 +58,6 @@ object KillBuildProcessTree {
             -1
         } finally {
             reader?.close()
-        }
-    }
-
-    fun addKillProcessTreeHook(projectId: String, buildId: String, vmSeqId: String) {
-        try {
-            Runtime.getRuntime().addShutdownHook(object : Thread() {
-                override fun run() {
-                    logger.info("start kill process tree")
-                    val killedProcessIds = killProcessTree(projectId, buildId, vmSeqId)
-                    logger.info("kill process tree done, ${killedProcessIds.size} process(s) killed, pid(s): $killedProcessIds")
-                }
-            })
-        } catch (t: Throwable) {
-            logger.warn("Fail to add shutdown hook", t)
         }
     }
