--- src/backend/ci/core/worker/worker-common/src/main/kotlin/com/tencent/devops/worker/common/api/utils/AgentBuildBaseInfoUtils.kt
@@ -1,63 +1,0 @@
-/*
- * Tencent is pleased to support the open source community by making BK-CI 蓝鲸持续集成平台 available.
- *
- * Copyright (C) 2019 THL A29 Limited, a Tencent company.  All rights reserved.
- *
- * BK-CI 蓝鲸持续集成平台 is licensed under the MIT license.
- *
- * A copy of the MIT License is included in this file.
- *
- *
- * Terms of the MIT License:
- * ---------------------------------------------------
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation
- * files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy,
- * modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the
- * Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT
- * LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN
- * NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
- * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package com.tencent.devops.worker.common.api.utils
-
-import com.tencent.devops.common.api.util.JsonUtil
-import com.tencent.devops.dispatch.pojo.thirdPartyAgent.AgentBuildBaseInfo
-import org.slf4j.LoggerFactory
-import java.util.Base64
-import kotlin.system.exitProcess
-
-object AgentBuildBaseInfoUtils {
-    private var buildInfo: AgentBuildBaseInfo? = null
-
-    fun setBuildInfo(info: AgentBuildBaseInfo) {
-        if (buildInfo != null) {
-            logger.info("Last build info is not null - $buildInfo")
-        }
-        buildInfo = info
-    }
-
-    fun getBuildInfo() = buildInfo
-
-    fun getBuildInfo(args: Array<String>): AgentBuildBaseInfo? {
-        if (args.isEmpty()) {
-            logger.error("Empty argument")
-            exitProcess(1)
-        }
-        val buildInfoStr = String(Base64.getDecoder().decode(args[0]))
-        try {
-            logger.info("Start read the build info ($buildInfoStr)")
-            return JsonUtil.to(buildInfoStr, AgentBuildBaseInfo::class.java)
-        } catch (t: Throwable) {
-            logger.warn("Fail to read the build Info", t)
-            exitProcess(1)
-        }
-    }
-
-    private val logger = LoggerFactory.getLogger(AgentBuildBaseInfoUtils::class.java)
-}
